Remove from mcpar-rfunc code specific to the food demand model.
The mc.setup function in an R model now returns a matrix giving the
recommended low and high values for the initial parameter guess.  The
RFunc class uses this matrix to infer the number of parameters and
makes the matrix values available to the rest of the code for use in
constructing an initial configuration.

The limits returned are recommended limits for the initial guesses,
but they are not hard limits on the parameters.  The MC algorithm can
and generally will stray outside these limits as it searches.  Therefore,
if there are hard limits (e.g., where the code will blow up if they are
violated), the model is still responsible for guarding them internally.