Accept remote proposals during burn-in.
Part of the purpose of the burn-in is to eliminate low-probability
transients at start-up.  The previous version of the code wasn't
effective at this because the markov chains weren't taking remote
proposals during burn-in.  Consequently, you wound up with a bunch of
low-likelihood samples in the output that had to be filtered.  This
change should fix that.